python-api/requirements.txt
flask==3.1.2
flask-cors==6.0.1
tensorflow==2.20.0
pillow==12.0.0
numpy==2.3.5
werkzeug==3.1.3